Rename PlainTextBadge to SecondaryInputBadge
in preparation for #1987
